requirements.txt
loguru
pulsectl
python-vlc
pygame
streamcontroller_plugin_tools